+++ ndk-sys/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Generate against upstream NDK build `11272921`. (#470)
 - Add `nativewindow` feature to link against `libnativewindow`. (#465)
 
 # 0.5.0 (2023-10-15)